+++ history/requirements.txt
@@ -1,2 +1,5 @@
 sqlalchemy
 psycopg2-binary
+pytest
+pytest-asyncio
+pytest-mock
