+++ history/tests/test_history.py
@@ -5,7 +5,7 @@ import pytest
 from sqlalchemy import select
 from sqlalchemy.orm import Session
 from event_schema.EventSchema import Event
-from app.history import HistoryPlayer
+from history_service.history import HistoryPlayer
 from tha_config import Config
